requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
psutil==5.9.7
pyautogui==0.9.54
pyperclip==1.8.2
Pillow==10.2.0
aiohttp==3.9.1
